Beta 7...
Add "bad_url" function and docos to PHP example.

Add --xrxcomments and --no-xrxcomments to docos.

Move "string.h" to "hdstring.h" to avoid name clashes on dumb compilers.

Change header/footer stuff (again) to check for paragraphs that
consist solely of whitespace, and to reset the Y position when
changing the top margin and the old Y was at the top of the form.


git-svn-id: http://svn.msweet.org/htmldoc@637 d512abc9-c3ee-0310-a016-ca57acef63f3